+++ Mesh_3/doc/Mesh_3/Concepts/MeshDomain_3.h
@@ -29,7 +29,7 @@ A segment, ray or line is said to intersect properly the domain boundary
 if it includes points which are strictly inside
 and strictly outside the domain (resp. the subdomain).
 
-\cgalHasModel `CGAL::Polyhedral_mesh_domain_3<Polyhedron,IGT,TriangleAccessor>`
+\cgalHasModel `CGAL::Polyhedral_mesh_domain_3<Polyhedron,IGT>`
 \cgalHasModel `CGAL::Labeled_mesh_domain_3<BGT>`
 
 \sa `MeshVertexBase_3`