+++ Mesh_3/include/CGAL/Polyhedral_mesh_domain_3.h
@@ -79,41 +79,71 @@ max_length(const Bbox_3& b)
 // -----------------------------------
 // Geometric traits generator
 // -----------------------------------
-template < typename Gt,
+template < typename GT,
            typename Use_exact_intersection_construction_tag >
 struct IGT_generator {};
 
-template < typename Gt >
-struct IGT_generator<Gt,CGAL::Tag_true>
+template < typename GT >
+struct IGT_generator<GT,CGAL::Tag_true>
 {
 #ifdef CGAL_MESH_3_NEW_ROBUST_INTERSECTION_TRAITS
-  typedef CGAL::Mesh_3::Robust_intersection_traits_3_new<Gt> type;
+  typedef CGAL::Mesh_3::Robust_intersection_traits_3_new<GT> type;
 #else // NOT CGAL_MESH_3_NEW_ROBUST_INTERSECTION_TRAITS
-  typedef CGAL::Mesh_3::Robust_intersection_traits_3<Gt> type;
+  typedef CGAL::Mesh_3::Robust_intersection_traits_3<GT> type;
 #endif // NOT CGAL_MESH_3_NEW_ROBUST_INTERSECTION_TRAITS
   typedef type Type;
 };
 
-template < typename Gt >
-struct IGT_generator<Gt,CGAL::Tag_false>
+template < typename GT >
+struct IGT_generator<GT,CGAL::Tag_false>
 {
-  typedef Gt type;
+  typedef GT type;
   typedef type Type;
 };
 
 }  // end namespace details
 }  // end namespace Mesh_3
 
-/**
- * @class Polyhedral_mesh_domain_3
- *
- *
- */
-template<class Polyhedron,/*FaceGraph*/
+
+/*!
+\ingroup PkgMesh3Domains
+
+The class `Polyhedral_mesh_domain_3` implements
+a domain defined by a simplicial polyhedral surface.
+
+The input polyhedral surface must be free of intersections.
+It must include (at least) one closed connected component
+that defines the boundary of the domain to be meshed.
+Inside this bounding component,
+the input polyhedral surface may contain
+several other components (closed or not)
+that will be represented in the final mesh.
+This class is a model of the concept `MeshDomain_3`.
+
+\tparam Polyhedron stands for the type of the input polyhedral surface(s),
+model of `FaceListGraph`.
+
+\tparam IGT stands for a geometric traits class
+providing the types and functors required to implement
+the intersection tests and intersection computations
+for polyhedral boundary surfaces. This parameter has to be instantiated
+with a model of the concept `IntersectionGeometricTraits_3`.
+
+\cgalModels `MeshDomain_3`
+
+\sa `IntersectionGeometricTraits_3`
+\sa `CGAL::make_mesh_3()`.
+*/
+#ifdef DOXYGEN_RUNNING
+template<class Polyhedron /*FaceGraph*/
+         ,class IGT>
+#else // DOXYGEN_RUNNING
+template<class Polyhedron, /*FaceGraph*/
          class IGT_,
-         class = CGAL::Default,
+         class TriangleAccessor = CGAL::Default,
          class Patch_id_ = void,
          class Use_exact_intersection_construction_tag = CGAL::Tag_true>
+#endif // DOXYGEN_RUNNING
 class Polyhedral_mesh_domain_3
 {
 public:
@@ -123,7 +153,7 @@ public:
 
   typedef Patch_id_ Patch_id;
 
-  /// Geometric object types
+  // Geometric object types
   typedef typename IGT::Point_3    Point_3;
   typedef typename IGT::Segment_3  Segment_3;
   typedef typename IGT::Ray_3      Ray_3;
@@ -134,24 +164,24 @@ public:
   //-------------------------------------------------------
   // Index Types
   //-------------------------------------------------------
-  /// Type of indexes for cells of the input complex
+  // Type of indexes for cells of the input complex
   typedef int Subdomain_index;
   typedef boost::optional<Subdomain_index> Subdomain;
 
-  /// Type of indexes for surface patch of the input complex
+  // Type of indexes for surface patch of the input complex
   typedef typename boost::property_map<Polyhedron,
                                        face_patch_id_t<Patch_id>
                                        >::type            Face_patch_id_pmap;
   typedef typename boost::property_traits<
     Face_patch_id_pmap>::value_type                       Surface_patch_index;
   typedef boost::optional<Surface_patch_index>            Surface_patch;
-  /// Type of indexes to characterize the lowest dimensional face of the input
-  /// complex on which a vertex lie
+
+  // Type of indexes to characterize the lowest dimensional face of the input
+  // complex on which a vertex lie
   typedef typename Mesh_3::internal::Index_generator<
     Subdomain_index, Surface_patch_index>::type           Index;
 
   typedef std::tuple<Point_3,Index,int> Intersection;
-
 
   typedef typename IGT::FT         FT;
 
@@ -197,8 +227,6 @@ public:
   typedef typename AABB_traits::Bounding_box              Bounding_box;
 
 public:
-
-  /// Default constructor
   Polyhedral_mesh_domain_3(CGAL::Random* p_rng = nullptr)
     : tree_()
     , bounding_tree_(&tree_)
@@ -206,27 +234,43 @@ public:
   {
   }
 
-  /**
-   * @brief Constructor. Construction from a polyhedral surface
-   * @param polyhedron the polyhedron describing the polyhedral surface
-   */
-  Polyhedral_mesh_domain_3(const Polyhedron& p,
-                           CGAL::Random* p_rng = nullptr)
+  /// \name Creation
+  /// @{
+
+  /*!
+    Construction from a bounding polyhedral surface which must be closed, and free of intersections.
+    The inside of `bounding_polyhedron` will be meshed.
+  */
+  Polyhedral_mesh_domain_3(const Polyhedron& bounding_polyhedron
+#ifndef DOXYGEN_RUNNING
+                           , CGAL::Random* p_rng = nullptr
+#endif
+                           )
     : tree_()
     , bounding_tree_(&tree_) // the bounding tree is tree_
     , p_rng_(p_rng)
   {
-    this->add_primitives(p);
-    if(! is_triangle_mesh(p)) {
+    this->add_primitives(bounding_polyhedron);
+    if(! is_triangle_mesh(bounding_polyhedron)) {
       std::cerr << "Your input polyhedron must be triangulated!\n";
       CGAL_error_msg("Your input polyhedron must be triangulated!");
     }
     this->build();
   }
 
+  /*!
+    Construction from a polyhedral surface, and a bounding polyhedral surface.
+    The first polyhedron must be entirely included inside `bounding_polyhedron`, which must be closed
+    and free of intersections.
+    Using this constructor enables to mesh a polyhedral surface which is not closed, or has holes.
+    The inside of `bounding_polyhedron` will be meshed.
+  */
   Polyhedral_mesh_domain_3(const Polyhedron& p,
-                           const Polyhedron& bounding_polyhedron,
-                           CGAL::Random* p_rng = nullptr)
+                           const Polyhedron& bounding_polyhedron
+#ifndef DOXYGEN_RUNNING
+                           , CGAL::Random* p_rng = nullptr
+#endif
+                           )
     : tree_()
     , bounding_tree_(new AABB_tree_)
     , p_rng_(p_rng)
@@ -241,22 +285,25 @@ public:
     this->build();
   }
 
-  /**
-   * Constructor.
-   *
+  /*!
    * Constructor from a sequence of polyhedral surfaces, and a bounding
    * polyhedral surface.
    *
-   * @param InputPolyhedraPtrIterator must an iterator of a sequence of
-   * pointers to polyhedra
+   * @tparam InputPolyhedraPtrIterator must be a model of
+   * `ForwardIterator` and value type `Polyhedron*`
    *
-   * @param bounding_polyhedron reference to the bounding surface
+   * @param begin iterator for a sequence of pointers to polyhedra
+   * @param end iterator for a sequence of pointers to polyhedra
+   * @param bounding_polyhedron the bounding surface
    */
   template <typename InputPolyhedraPtrIterator>
   Polyhedral_mesh_domain_3(InputPolyhedraPtrIterator begin,
                            InputPolyhedraPtrIterator end,
-                           const Polyhedron& bounding_polyhedron,
-                           CGAL::Random* p_rng = nullptr)
+                           const Polyhedron& bounding_polyhedron
+#ifndef DOXYGEN_RUNNING
+                           , CGAL::Random* p_rng = nullptr
+#endif
+                           )
     : p_rng_(p_rng)
     , delete_rng_(false)
   {
@@ -274,20 +321,24 @@ public:
     this->build();
   }
 
-  /**
-   * Constructor.
-   *
-   * Constructor from a sequence of polyhedral surfaces, without bounding
-   * surface. The domain will always answer false to "is_in_domain"
+  /*!
+   * Constructor from a sequence of polyhedral surfaces, without a bounding
+   * surface. The domain will always answer `false` to `is_in_domain()`
    * queries.
    *
-   * @param InputPolyhedraPtrIterator must an iterator of a sequence of
-   * pointers to polyhedra
+   * @tparam InputPolyhedraPtrIterator must be a model of
+   * `ForwardIterator` and value type `Polyhedron*`
+   *
+   * @param begin iterator for a sequence of pointers to polyhedra
+   * @param end iterator for a sequence of pointers to polyhedra
    */
   template <typename InputPolyhedraPtrIterator>
-  Polyhedral_mesh_domain_3(InputPolyhedraPtrIterator begin,
-                           InputPolyhedraPtrIterator end,
-                           CGAL::Random* p_rng = nullptr)
+  Polyhedral_mesh_domain_3(InputPolyhedraPtrIterator begin
+                           , InputPolyhedraPtrIterator end
+#ifndef DOXYGEN_RUNNING
+                           , CGAL::Random* p_rng = nullptr
+#endif
+                           )
     : p_rng_(p_rng)
   {
     if(begin != end) {
@@ -299,7 +350,9 @@ public:
     bounding_tree_ = 0;
   }
 
-  /// Destructor
+  /// @}
+
+  // Destructor
   ~Polyhedral_mesh_domain_3() {
     if(bounding_tree_ != 0 && bounding_tree_ != &tree_) {
       delete bounding_tree_;
@@ -311,9 +364,9 @@ public:
   }
 
   /**
-   * Constructs  a set of \ccc{n} points on the surface, and output them to
-   *  the output iterator \ccc{pts} whose value type is required to be
-   *  \ccc{std::pair<Points_3, Index>}.
+   * constructs a set of `n` points on the surface, and output them to
+   *  the output iterator `pts` whose value type is required to be
+   *  `std::pair<Points_3, Index>`.
    */
   struct Construct_initial_points
   {
@@ -342,9 +395,9 @@ public:
 
 
   /**
-   * Returns true if point~\ccc{p} is in the domain. If \ccc{p} is in the
+   * Returns true if point `p` is in the domain. If `p` is in the
    *  domain, the parameter index is set to the index of the subdomain
-   *  including $p$. It is set to the default value otherwise.
+   *  including `p`. It is set to the default value otherwise.
    */
   struct Is_in_domain
   {
@@ -363,16 +416,16 @@ public:
     return tree_.closest_point(p);
   }
 
-  /// Allowed query types
+  // Allowed query types
   typedef boost::mpl::vector<Segment_3, Ray_3, Line_3> Allowed_query_types;
 
   /**
-   * Returns true is the element \ccc{type} intersect properly any of the
-   * surface patches describing the either the domain boundary or some
+   * Returns `true` if the element `type` intersects properly any of the
+   * surface patches describing either the domain boundary or some
    * subdomain boundary.
-   * \ccc{Type} is either \ccc{Segment_3}, \ccc{Ray_3} or \ccc{Line_3}.
+   * `Type` is either `Segment_3`, `Ray_3` or `Line_3`.
    * Parameter index is set to the index of the intersected surface patch
-   * if \ccc{true} is returned and to the default \ccc{Surface_patch_index}
+   * if `true` is returned and to the default `Surface_patch_index`
    * value otherwise.
    */
   struct Do_intersect_surface
@@ -408,12 +461,12 @@ public:
   }
 
   /**
-   * Returns a point in the intersection of the primitive \ccc{type}
+   * Returns a point in the intersection of the primitive `type`
    * with some boundary surface.
-   * \ccc{Type1} is either \ccc{Segment_3}, \ccc{Ray_3} or \ccc{Line_3}.
-   * The integer \ccc{dimension} is set to the dimension of the lowest
+   * `Type1` is either `Segment_3`, `Ray_3` or `Line_3`.
+   * The integer `dimension` is set to the dimension of the lowest
    * dimensional face in the input complex containing the returned point, and
-   * \ccc{index} is set to the index to be stored at a mesh vertex lying
+   * `index` is set to the index to be stored at a mesh vertex lying
    * on this face.
    */
   struct Construct_intersection